Factor out keyGeomToLayout
This can be re-used and tested when outside of line function
